Added Delete food functionality
- You can now delete foods.
- Stylesheet changes
- List View changes for remove link
- If the user does not have foods, a prompt now appears telling the
user how to create a food.
- Food Items controller functions added, destroy is for removing foods,
new_food and remove are for the modal buttons. I will see if new_page
and remove_page are javascript.erb files for the modals.
—————————————————————
- There is a modal for deleting foods but I may get rid of it since i
cannot seem to get deleting multiple foods at once to work, which was
the purpose of the modal. I’m going to keep it for now but if I cannot
get it to work by thursday I am removing it.